requirements.txt
langgraph>=0.0.45
langchain>=0.3.0
langchain-groq>=0.1.4
groq>=0.5.0
python-telegram-bot>=21.4
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.30.6
sqlite-utils>=3.36.0
langchain-community>=0.3.0
langgraph>=0.0.45
langchain>=0.3.0
langchain-groq>=0.1.4
groq>=0.5.0
python-telegram-bot>=21.4
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.30.6
python-telegram-bot==21.4
pyTelegramBotAPI
langgraph_groq_agent